DataBase/requirements.txt
psycopg2
sqlalchemy
python-dotenv
fastapi
pydantic
uvicorn